Make LinkedStudy constituent keys ordered
By making the two parts of the composite keys ordered it reduces the
database retrieval for finding whether two study ids are linked. This
also simplifies the implementation of service methods.

Also add CHECK constraints to the linked_study table to make sure that
the constraint is also maintained at the database level.